aggregator: Move create_nonce_streams to rpc crate.
